+++ CHANGELOG.md
@@ -1,13 +1,22 @@
+# 1.2.2 
+
+### Implemented a robust error-handling mechanism.  
+
+- `AppException` handler to automatically map specific errors
+- Added `ErrorMapper` to map your own errors
+- Ensured consistent logging for easier debugging.
+- Enhanced Flutter error reporting
+
 # 1.2.1
 
-### Improved Platform Support [2025-03-21]  
+### Improved Platform Support 
 
 - Ensured proper platform support, including explicit web support  
 - Added theme image and improved pub.dev visibility  
 
 # 1.2.0  
 
-### Soul Theme Engine [2025-03-21]  
+### Soul Theme Engine 
 
 - Introduced **Soul Theme Engine**, an advanced theming solution  
 - Added **10+ premade themes** for easy customization  
@@ -15,14 +24,14 @@
 
 # 1.1.0  
 
-### Theme Engine [2025-03-20]  
+### Theme Engine  
 
 - Allowing for structured theming with color palettes  
 - Added **Currency Format**  
 
 # 1.0.0  
 
-### Initial Release [2025-03-19]  
+### Initial Release 
 
 - Released the initial stable version of Flutter Addons  
 - Core functionality implemented